Pulled recipe card into separate component
